requirements.txt
gymnasium[accept-rom-license]
ale-py
torch
numpy
opencv-python
matplotlib
